ttl is a 32bit field

+++ main.c
@@ -143,7 +143,7 @@ struct ResourceRecord {
   char *name;
   uint16_t type;
   uint16_t class;
-  uint16_t ttl;
+  uint32_t ttl;
   uint16_t rd_length;
   union ResourceData rd_data;
   struct ResourceRecord* next; // for linked list
